+++ app/models/customer.rb
@@ -10,4 +10,5 @@ class Customer < ApplicationRecord
   update_index('customers#customer', :self)
   has_paper_trail(meta: { store_id: :store_id })
   scope :by_store, ->(store_id) { where(store_id: store_id) }
+  self.per_page = 10
 end